Work around mapbox-gl-native/15574. Avoid re-setting content inset to safeArea. Panning fix.
Fix CarPlay puck positioning, panning away from navigation, navigation -> overview transitions.

The workaround should be future proof as the https://github.com/mapbox/mapbox-gl-native/issues/15574 exists only with MGLMapView.contentInset.
Additionally, it simplifies the code as it doesn't need to ensure that safeArea and contentInset are kept in sync: according to work in https://github.com/mapbox/mapbox-navigation-ios/pull/2211, contentInset is to be used to position puck and this causes map position sudden change when panning.

Fixes: https://github.com/mapbox/mapbox-navigation-ios/issues/2190